refactor(lspconfig): use variable for ensure_installed
Switch to using the ensure_installed variable instead of a hardcoded
elixirls version in mason-lspconfig setup. This change improves
maintainability and flexibility, allowing the list of LSP servers to be
managed elsewhere and updated more easily.

Previously, elixirls was pinned to v0.29.3 due to issues with newer
versions. By referencing the variable, future updates to the list or
version can be made without modifying this section directly.